# perf: Optimize and simplify sender recovery