… crates/stages/src/stages/stream.rs