Stumbled on this and took a better look, did some optimization and it simplifies the process.

This is a more straightforward way to do result reporting from tasks. It created N amount of channels that every task pushes results to, and with it, it does not require additional sorting of transactions ids.

And Boxed the Error as it is probably bigger than our result.