@rakita 2023-03-22
New way
```
draganrakita@192 stages % cargo bench --package reth-stages --bench criterion --features test-utils -- Stages/SendersRecovery-batch-1000
   Compiling reth-stages v0.1.0 (/Users/draganrakita/workspace/reth/crates/stages)
    Finished bench [optimized] target(s) in 4.35s
     Running benches/criterion.rs (/Users/draganrakita/workspace/reth/target/release/deps/criterion-730c9f55092a1b94)
Benchmarking Stages/SendersRecovery-batch-1000: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 48.0s.
Benchmarking Stages/SendersRecovery-batch-1000: Collecting 10 samples in estimated 47.951 s (10 iterations)

Stages/SendersRecovery-batch-1000
                        time:   [3.9731 s 4.0102 s 4.0482 s]
                        change: [-1.2876% -0.0502% +1.3153%] (p = 0.94 > 0.05)
                        No change in performance detected.
Benchmarking Stages/SendersRecovery-batch-10000: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 48.7s.
Stages/SendersRecovery-batch-10000
                        time:   [4.0837 s 4.1036 s 4.1291 s]
                        change: [-4.0650% -2.1451% -0.2816%] (p = 0.05 > 0.05)
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
Benchmarking Stages/SendersRecovery-batch-100000: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 50.0s.
Stages/SendersRecovery-batch-100000
                        time:   [4.0880 s 4.1062 s 4.1245 s]
                        change: [-1.2251% +0.8742% +2.7287%] (p = 0.42 > 0.05)
                        No change in performance detected.
```

old way:
```
draganrakita@192 stages % cargo bench --package reth-stages --bench criterion --features test-utils -- Stages/SendersRecovery-batch-1000
   Compiling reth-stages v0.1.0 (/Users/draganrakita/workspace/reth/crates/stages)
    Finished bench [optimized] target(s) in 4.60s
     Running benches/criterion.rs (/Users/draganrakita/workspace/reth/target/release/deps/criterion-730c9f55092a1b94)
Benchmarking Stages/SendersRecovery-batch-1000: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 51.5s.
Stages/SendersRecovery-batch-1000
                        time:   [4.2869 s 4.2928 s 4.2989 s]
                        change: [+6.0266% +7.0460% +8.0610%] (p = 0.00 < 0.05)
                        Performance has regressed.
Benchmarking Stages/SendersRecovery-batch-10000: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 51.7s.
Stages/SendersRecovery-batch-10000
                        time:   [4.2682 s 4.2732 s 4.2781 s]
                        change: [+3.4929% +4.1349% +4.6667%] (p = 0.00 < 0.05)
                        Performance has regressed.
Benchmarking Stages/SendersRecovery-batch-100000: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 51.5s.
Stages/SendersRecovery-batch-100000
                        time:   [4.2733 s 4.2941 s 4.3217 s]
                        change: [+3.8634% +4.5779% +5.3469%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 3 outliers among 10 measurements (30.00%)
  1 (10.00%) low mild
  2 (20.00%) high severe
```

Not that big of an impact, 4-5% faster